+++ CHANGELOG.md
@@ -24,6 +24,7 @@ When the Unreleased section becomes a new version, duplicate the Template to cre
 ## [Unreleased]
 ### Added
 - Optional `theme` parameter to override light/dark theme for authorization UX.
+- `address` and `birthdate` scopes
 ### Changed
 ### Removed
 ### Deprecated